Rename data attributes
Oh and fix the background image disabling, I guess I didn't notice that not working.
Oh yeah, that's why I had picked Unsounded for an example comic before, because it has a bunch of good background image elements (including integrated with the page content). (In addition to the nice art and humour.)