Only initialize socket when enabled for a site
This makes it more efficient (for sites where it's disabled), and avoids an error in Firefox:

    Firefox can’t establish a connection to the server at ws://localhost:4284/socket.io/?EIO=3&transport=websocket.
    The connection to ws://localhost:4284/socket.io/?EIO=3&transport=websocket was interrupted while the page was loading.

It ALMOST works in Firefox now, but the binary data messages are not working. If I disable sending the result_array_buffer on the server, the client receives the "superrez-result" event, otherwise it doesn't.